requirements.txt
alembic==1.16.5
apispec==6.8.4
apispec-webframeworks==1.2.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==5.21.0
greenlet==3.2.4
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.0.1
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
PyYAML==6.0.3
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
